refactor: revert back to building config in build function rather than just in CLI
